aplicado SRP do SOLID em algumas classes
