+++ src/main/java/com/github/fsousa1987/effecti/testbackend/domain/service/BiddingService.java
@@ -2,7 +2,7 @@ package com.github.fsousa1987.effecti.testbackend.domain.service;
 
 import com.github.fsousa1987.effecti.testbackend.api.response.BatchBidResponse;
 
-public interface BindingService {
+public interface BiddingService {
 
     BatchBidResponse searchBidsFromWeb();
